A corriger

+++ ft_memcpy.c
@@ -6,17 +6,18 @@
 /*   By: vcarbone <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/14 15:52:51 by vcarbone          #+#    #+#             */
-/*   Updated: 2016/11/14 18:58:45 by vcarbone         ###   ########.fr       */
+/*   Updated: 2016/11/16 16:31:07 by vcarbone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <string.h>
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char		*d;
 	const char	*s;
-	int			i;
+	size_t		i;
 
 	d = dst;
 	s = src;
